+++ src/registry-auth-locator/creds-store.ts
@@ -1,0 +1,13 @@
+import { CredentialProvider } from "./credential-provider";
+import { DockerConfig } from "./types";
+
+export class CredsStore extends CredentialProvider {
+  isApplicable(registry: string, dockerConfig: DockerConfig): boolean {
+    return dockerConfig.credsStore !== undefined && dockerConfig.credsStore.length > 0;
+  }
+
+  getCredentialProviderName(registry: string, dockerConfig: DockerConfig): string {
+    // @ts-ignore
+    return dockerConfig.credsStore;
+  }
+}
